backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn>=0.27.0

# JWT handling
python-jose[cryptography]>=3.3.0

# Configuration management
pydantic-settings>=2.0.0

# Database
sqlmodel>=0.0.16
asyncpg>=0.29.0
greenlet>=3.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
aiosqlite>=0.19.0
